feat(ThreeColumnLayout): Make gap property be set through useContext
90% of the uses for the ThreeColumnLayout are where the gap between the
three columns is the same amount, so it would be easier to set a gap
value in the parent element and have all of the children automatically
use the gap value that was passed to the parent, instead of manually
inputting gap values for the left, middle and right column.

If I do need to set explicity gap values for one specific column, I can
just use an optional prop and have that prop's value override the
previous prop value that was derrived from the parent element.