chore: Move starlight folder out of root folder
I had an astro starlight project nested in the folder of a React project,
which didn't allow me to deploy the astro and React project independently.
I decided to move all of the the code in the starlight folder to its own
folder so the React and astro project can be their own Git repos I can
use package dependencies to make code in the React project a local
dependency of the astro project.